Move Text and Dictionary to types.py
