+++ lib/__init__.py
@@ -6,8 +6,8 @@ from .io import load_json, save_json
 from .merge import merge
 from .servant import servant_dict, servant_list, skill_dict
 from .sound import Sound, sound_list
-from .text import Dictionary, Text, load_item_dictionary
-from .types import Item, Servant
+from .text import load_item_dictionary
+from .types import Dictionary, Item, Servant, Text
 from .validate import validate_servant, validate_servants
 
 logging.getLogger(__package__).addHandler(logging.NullHandler())
